scripts/requirements-sync.txt
httpx>=0.24.0
python-dotenv>=0.19.0
supabase>=2.0.0
asyncio>=3.4.3
